Chaging more things to fit paradigms, trying initial behaiour and setting scripts to right table
