+++ checkersUnity/Assets/Prefabs/Board.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 795c91961569e864fb8aa4942fa51435
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
