+++ checkersUnity/Assets/Scripts/Controllers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f57f5d0b887e68b4bb39335ffd71b480
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
