+++ checkersUnity/Assets/Scripts/Controllers/TableConstructor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bbe7b83e8725a644fb898e28e9826132
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
