+++ checkersUnity/Assets/Scripts/Objects/BoardPiece.cs
