+++ checkersUnity/Assets/Scripts/Objects/Piece.cs
